+++ apps/web/src/pages/AccountPage/AccountPage.tsx
@@ -1,14 +1,10 @@
-import styled from "@emotion/styled";
-
-const Page = styled.div`
-  padding: 2rem;
-`;
+import * as S from "./AccountPage.styled";
 
 export function AccountPage() {
   return (
-    <Page>
+    <S.Page>
       <h1>계정관리</h1>
       <p>계정 정보를 관리합니다.</p>
-    </Page>
+    </S.Page>
   );
 }
