+++ apps/web/src/pages/AccountPage/index.ts
@@ -1,0 +1,1 @@
+export { AccountPage } from "./AccountPage";
