+++ apps/web/src/pages/LoginPage/LoginPage.tsx
@@ -1,14 +1,10 @@
-import styled from "@emotion/styled";
-
-const Page = styled.div`
-  padding: 2rem;
-`;
+import * as S from "./LoginPage.styled";
 
 export function LoginPage() {
   return (
-    <Page>
+    <S.Page>
       <h1>로그인</h1>
       <p>계정에 로그인합니다.</p>
-    </Page>
+    </S.Page>
   );
 }
