+++ apps/web/src/pages/MyGardenPage/MyGardenPage.tsx
@@ -1,14 +1,10 @@
-import styled from "@emotion/styled";
-
-const Page = styled.div`
-  padding: 2rem;
-`;
+import * as S from "./MyGardenPage.styled";
 
 export function MyGardenPage() {
   return (
-    <Page>
+    <S.Page>
       <h1>나의 정원</h1>
       <p>블로그 활동을 잔디처럼 시각화한 공간입니다.</p>
-    </Page>
+    </S.Page>
   );
 }
