+++ apps/web/src/pages/MyGardenPage/index.ts
@@ -1,0 +1,1 @@
+export { MyGardenPage } from "./MyGardenPage";
