+++ apps/web/src/pages/SettingsPage/SettingsPage.styled.ts
@@ -1,0 +1,5 @@
+import styled from "@emotion/styled";
+
+export const Page = styled.div`
+  padding: 2rem;
+`;
