+++ apps/web/src/pages/SettingsPage/SettingsPage.tsx
@@ -1,14 +1,10 @@
-import styled from "@emotion/styled";
-
-const Page = styled.div`
-  padding: 2rem;
-`;
+import * as S from "./SettingsPage.styled";
 
 export function SettingsPage() {
   return (
-    <Page>
+    <S.Page>
       <h1>설정</h1>
       <p>서비스 설정을 변경합니다.</p>
-    </Page>
+    </S.Page>
   );
 }
