+++ apps/web/src/pages/SettingsPage/index.ts
@@ -1,0 +1,1 @@
+export { SettingsPage } from "./SettingsPage";
