+++ apps/web/src/pages/WidgetPage/WidgetPage.tsx
@@ -1,14 +1,10 @@
-import styled from "@emotion/styled";
-
-const Page = styled.div`
-  padding: 2rem;
-`;
+import * as S from "./WidgetPage.styled";
 
 export function WidgetPage() {
   return (
-    <Page>
+    <S.Page>
       <h1>위젯</h1>
       <p>블로그에 삽입할 위젯을 관리합니다.</p>
-    </Page>
+    </S.Page>
   );
 }
