+++ apps/web/src/pages/WidgetPage/index.ts
@@ -1,0 +1,1 @@
+export { WidgetPage } from "./WidgetPage";
